# documentation/static/openapi/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Provider-neutral HTTP and server-sent-events API for running agent turns, reading retained messages, resolving tool decisions, and managing subagents.",
        "title": "AgentCLI HTTP API",
        "contact": {
            "name": "AgentCLI API support"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/license/mit"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/sessions/{mainAgentSessionID}/subagent-confirmations": {
            "get": {
                "description": "Returns durable pending subagent confirmation requests so clients can recover after attaching late or reconnecting.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Confirmations"
                ],
                "summary": "List pending confirmations for a main agent session",
                "operationId": "listPendingSubagentConfirmations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Main agent session ID",
                        "name": "mainAgentSessionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agentcli.PendingSubagentConfirmationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/sessions/{mainAgentSessionID}/subagent-permissions": {
            "get": {
                "description": "Returns durable pending subagent permission requests so clients can recover after attaching late or reconnecting.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "List pending permissions for a main agent session",
                "operationId": "listPendingSubagentPermissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Main agent session ID",
                        "name": "mainAgentSessionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agentcli.PendingSubagentPermissionsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/sessions/{mainAgentSessionID}/subagents": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subagents"
                ],
                "summary": "List subagents owned by a main agent session",
                "operationId": "listSubagents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owning main agent session ID",
                        "name": "mainAgentSessionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Include closed subagent records",
                        "name": "include_closed",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agentcli.SubagentsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subagents"
                ],
                "summary": "Create and asynchronously start a subagent",
                "operationId": "createSubagent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owning main agent session ID",
                        "name": "mainAgentSessionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Subagent definition and initial message",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/agentcli.CreateSubagentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/agentcli.SubagentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/sessions/{mainAgentSessionID}/subagents/{subagentID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subagents"
                ],
                "summary": "Read one owned subagent",
                "operationId": "getSubagent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owning main agent session ID",
                        "name": "mainAgentSessionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Subagent ID",
                        "name": "subagentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agentcli.SubagentResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Destructively closes one subagent, interrupting active work, dropping queued input, and cancelling outstanding response-scope result obligations when necessary, while retaining transcript and completed event history. Bind this endpoint to an explicit user action. A successful close also emits subagent_closed on the main agent session event stream.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subagents"
                ],
                "summary": "Close one owned subagent",
                "operationId": "closeSubagent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owning main agent session ID",
                        "name": "mainAgentSessionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Subagent ID",
                        "name": "subagentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agentcli.SubagentResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "agentcli.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "agentcli.APIErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/agentcli.APIError"
                }
            }
        },
...
        "agentcli.CreateSubagentRequest": {
            "type": "object",
            "required": [
                "message",
                "name"
            ],
            "properties": {
                "label": {
                    "type": "string"
                },
                "main_agent_turn_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "agentcli.PendingSubagentConfirmationsResponse": {
            "type": "object",
            "properties": {
                "confirmations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/agentcli.SubagentConfirmationReference"
                    }
                }
            }
        },
        "agentcli.PendingSubagentPermissionsResponse": {
            "type": "object",
            "properties": {
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/agentcli.SubagentPermissionReference"
                    }
                }
            }
        },
...
        "agentcli.SubagentsResponse": {
            "type": "object",
            "properties": {
                "subagents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/agentcli.SubagentResponse"
                    }
                }
            }
        },
...
    }
}